// src/app/main/reports/report-type.ts
export type DurationType = 'month' | 'week' | 'day';
// ...
export type CoilSummary = {
  coil: Coil;
  usages: CoilUsagesEntity[];
};
export type Coil = {
  documentID: string;
  coilId: string;
  materialCode: string;
  description: string;
  markedComplete: boolean;
  lengthStartFt: number;
  lengthUsedFt: number;
  lengthNonExemptScrapFt: number;
  lengthExemptScrapFt: number;
  lengthOtherAdjustmentsFt: number;
  weightStartLbs: number;
  dateIn: string;
  dateOut?: string | null;
  importDate: string;
  vendorName: string;
  vendorCode: string;
  heatNumber: string;
  purchaseOrder: string;
  user1: string;
  user2: string;
  storeLocation: string;
};
export type CoilUsagesEntity = {
  orderCode: string;
  customerName: string;
  machineNumber: number;
  date: string;
  goodFt: number;
  scrapFt: number;
  otherFt: number;
  netChangeFt: number;
};

export type FilterState = {
  startDate: moment.Moment;
  endDate: moment.Moment;
  duration?: DurationType;
  group1?: string;
  group2?: string;
  machines?: number[];
  downtimeCodes?: number[];
  delayCodes?: number[];
  shifts?: number[];
  scheduleStatus?: string;
  isBack?: boolean;
};